B: split runtime menu-input plumbing into runtime/input.rs
